pyproject.toml
[project]
name = "some-python-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Pillow>=10.0.0",
    "click>=8.0.0"
]
